updated the models to have float instead of numeric to avoid issue

+++ python-project-starter/app/models/portfolio_value.py
@@ -4,7 +4,7 @@ class PortfolioValue(db.Model):
     __tablename__ = "PortfolioValues"
 
     id = db.Column(db.Integer, primary_key=True)
-    value = db.Column(db.Numeric(10, 2), nullable=False)
+    value = db.Column(db.Float, nullable=False)
     date = db.Column(db.DateTime(timezone=True), server_default=db.func.now(), nullable=False)
     user_id = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False)
